services/graphql-api/pyproject.toml
[project]
name = "graphql-api"
version = "0.1.0"
description = "GraphQL API serving layer for the DE capstone lakehouse"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0,<1.0",
    "uvicorn[standard]>=0.27.0,<1.0",
    "strawberry-graphql[fastapi]>=0.220.0,<1.0",
    "duckdb>=1.0,<2.0",
    "pyiceberg[s3fs]>=0.7,<1.0",
    "pyarrow>=14.0,<18.0",
    "aiokafka>=0.10.0,<1.0",
    "PyJWT>=2.8.0,<3.0",
    "pydantic-settings>=2.1,<3.0",
    "aiosqlite>=0.20.0,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
